// src/enums/EnumsHSFormula.ts
export enum EnumNameField {
    Строка = 'Строка',
    Число = 'Число',
    Массив = 'Массив'
}
export enum EnumTypeField {
    String = 'string',
    Number = 'number',
    Array = 'array'
}
export enum EnumCategoriesFormula {
    Array = 'Array',
    Information = 'Information',
    Financial = 'Financial',
    Engineering = 'Engineering',
    Logical = 'Logical',
    MathAndTrigonometry = 'MathAndTrigonometry',
    Operator = 'Operator',
    Statistical = 'Statistical',
    Text = 'Text'
}

export enum EnumCategoriesRusFormula {
    Все = 'Все',
    Текстовые = 'Текстовые',
    Статистические = 'Статистические',
    ДатаИВремя = 'Дата и время',
    Инженерные = 'Инженерные',
    БазыДанных = 'Базы данных',
    Финансовые = 'Финансовые',
    Математические = 'Математические',
    ПоискИСсылки = 'Поиск и ссылки',
    Информационные = 'Информационные',
    Логические = 'Логические'
}

export enum EnumNamesFormula {
    TEST = 'TEST',
    ABS = 'ABS',
    СУММ = 'СУММ',
    ПРОИЗВЕД = 'ПРОИЗВЕД',
    ЧАСТНОЕ = 'ЧАСТНОЕ',
    СУММЕСЛИ = 'СУММЕСЛИ',
    ОКРУГЛ = 'ОКРУГЛ',
    ОКРУГЛВВЕРХ = 'ОКРУГЛВВЕРХ',
    ОКРУГЛВНИЗ = 'ОКРУГЛВНИЗ',
    ОТБР = 'ОТБР',
    ЦЕЛОЕ = 'ЦЕЛОЕ',
    КОРЕНЬ = 'КОРЕНЬ',
    СТЕПЕНЬ = 'СТЕПЕНЬ',
    ДЛСТР = 'ДЛСТР',
    СРЗНАЧ = 'СРЗНАЧ',
    МИН = 'МИН',
    МАКС = 'МАКС',
    ЕСЛИ = 'ЕСЛИ',
    ИСТИНА = 'ИСТИНА',
    ЛОЖЬ = 'ЛОЖЬ',
    И = 'И',
    НЕ= 'НЕ',
    ИЛИ= 'ИЛИ',
    СЧЁТ='СЧЁТ',
    ПРАВСИМВ='ПРАВСИМВ',
    ЛЕВСИМВ='ЛЕВСИМВ',
}  

export enum EnumNamesEngFormula {
    TEST = 'TEST',
    AND = 'AND',
    ABS = 'ABS',
    SUM = 'SUM',
    PRODUCT = 'PRODUCT',
    QUOTIENT = 'QUOTIENT',
    SUMIF = 'SUMIF',
    ROUND = 'ROUND',
    ROUNDUP = 'ROUNDUP',
    ROUNDDOWN = 'ROUNDDOWN',
    TRUNC = 'TRUNC',
    INT = 'INT',
    SQRT = 'SQRT',
    POWER = 'POWER',
    LEN = 'LEN',
    AVERAGE = 'AVERAGE',
    MAX = 'MAX',
    MIN = 'MIN',
    IF = 'IF',
    FALSE = 'FALSE',
    TRUE = 'TRUE',
    NOT = 'NOT',
    OR = 'OR',
    COUNT = 'COUNT',
    RIGHT='RIGHT',
    LEFT='LEFT',
} 